+++ src/Diagnostics.DataProviders/K8SELogAnalyticsClient.cs
@@ -14,59 +14,20 @@ namespace Diagnostics.DataProviders
     {
         private K8SELogAnalyticsDataProviderConfiguration _configuration;
 
-        private K8SETokenService _tokenService;
-
-        /*private string workspaceId;
-        private string clientId;
-        private string clientSecret;
-
-        private string domain;
-        private string authEndpoint;
-        private string tokenAudience;
-
-        private ActiveDirectoryServiceSettings adSettings;
-        private Microsoft.Rest.ServiceClientCredentials creds;*/
-
         public K8SELogAnalyticsClient(K8SELogAnalyticsDataProviderConfiguration configuration)
         {
             _configuration = configuration;
-            K8SETokenService.Instance.Initialize(_configuration);
-
-            /*workspaceId = _configuration.WorkspaceId;
-            clientId = _configuration.ClientId;
-            clientSecret = _configuration.ClientSecret;
-
-            domain = _configuration.Domain;
-            authEndpoint = _configuration.AuthEndpoint;
-            tokenAudience = _configuration.TokenAudience;*/
-
+            K8SELogAnalyticsTokenService.Instance.Initialize(_configuration);
         }
 
         public async Task<DataTable> ExecuteQueryAsync(string query)
         {
-            /*creds = await ApplicationTokenProvider.LoginSilentAsync(domain, clientId, clientSecret, adSettings);
-
-            adSettings = new ActiveDirectoryServiceSettings
-            {
-                AuthenticationEndpoint = new Uri(authEndpoint),
-                TokenAudience = new Uri(tokenAudience),
-                ValidateAuthority = true
-            };*/
-            /*var client = new OperationalInsightsDataClient(_tokenService.GetCreds());
-            client.WorkspaceId = _tokenService.GetWorkspaceId();*/
-
-            var queryResults = await K8SETokenService.Instance.GetClient().QueryAsync(query);
+            var queryResults = await K8SELogAnalyticsTokenService.Instance.GetClient().QueryAsync(query);
 
             var dataTable = ResultAsDataTable(queryResults);
 
             return dataTable;
         }
-
-        //assign creds asynchronously
-        /*private async Task GetCredsAsync()
-        {
-            creds = await ApplicationTokenProvider.LoginSilentAsync(domain, clientId, clientSecret, adSettings);
-        }*/
 
         //convert results to DataTable
         private DataTable ResultAsDataTable(Microsoft.Azure.OperationalInsights.Models.QueryResults results)
