+++ src/Diagnostics.DataProviders/TokenService/K8SELogAnalyticsTokenService.cs
@@ -1,30 +1,54 @@
-﻿using Microsoft.IdentityModel.Clients.ActiveDirectory;
+﻿using Diagnostics.DataProviders.DataProviderConfigurations;
+using Diagnostics.Logger;
+using Microsoft.Azure.OperationalInsights;
+using Microsoft.Rest.Azure.Authentication;
 using System;
 using System.Collections.Generic;
 using System.Text;
-using Diagnostics.DataProviders.DataProviderConfigurations;
+using System.Threading.Tasks;
 
 namespace Diagnostics.DataProviders.TokenService
 {
-    public class K8SELogAnalyticsTokenService : TokenServiceBase
+    class K8SELogAnalyticsTokenService : LogAnalyticsTokenServiceBase
     {
         private static readonly Lazy<K8SELogAnalyticsTokenService> instance = new Lazy<K8SELogAnalyticsTokenService>(() => new K8SELogAnalyticsTokenService());
-
         public static K8SELogAnalyticsTokenService Instance => instance.Value;
-        protected override AuthenticationContext AuthenticationContext { get; set; }
-        protected override ClientCredential ClientCredential { get; set; }
-        protected override string Resource { get; set; }
+
+        protected override string workspaceId { get; set; }
+        protected override string clientId { get; set; }
+        protected override string clientSecret { get; set; }
+
+        protected override string domain { get; set; }
+        protected override string authEndpoint { get; set; }
+        protected override string tokenAudience { get; set; }
+
+        protected override ActiveDirectoryServiceSettings adSettings { get; set; }
+        protected override Microsoft.Rest.ServiceClientCredentials creds { get; set; }
+
+        protected override OperationalInsightsDataClient client { get; set; }
+
+
         protected override string TokenServiceName { get; set; }
+
+        public OperationalInsightsDataClient GetClient()
+        {
+            return client;
+        }
 
         public void Initialize(K8SELogAnalyticsDataProviderConfiguration k8SELogAnalyticsDataProviderConfiguration)
         {
-            Resource = k8SELogAnalyticsDataProviderConfiguration.WorkspaceId;
-            AuthenticationContext = new AuthenticationContext("https://login.microsoftonline.com/microsoft.onmicrosoft.com");
-            //AuthenticationContext = new AuthenticationContext(k8SELogAnalyticsDataProviderConfiguration.Domain);
-            ClientCredential = new ClientCredential(k8SELogAnalyticsDataProviderConfiguration.ClientId,
-                                                    k8SELogAnalyticsDataProviderConfiguration.ClientSecret);
+            workspaceId = k8SELogAnalyticsDataProviderConfiguration.WorkspaceId;
+            clientId = k8SELogAnalyticsDataProviderConfiguration.ClientId;
+            clientSecret = k8SELogAnalyticsDataProviderConfiguration.ClientSecret;
+
+            domain = k8SELogAnalyticsDataProviderConfiguration.Domain;
+            authEndpoint = k8SELogAnalyticsDataProviderConfiguration.AuthEndpoint;
+            tokenAudience = k8SELogAnalyticsDataProviderConfiguration.TokenAudience;
+
             TokenServiceName = "K8SELogAnalyticsTokenRefresh";
+
             StartTokenRefresh();
         }
+        
     }
 }
